Merge 1.2.0 developments into master

+++ iam-login-service/src/test/java/it/infn/mw/iam/test/api/tokens/AccessTokenGetListTests.java
@@ -1,24 +1,13 @@
 package it.infn.mw.iam.test.api.tokens;
 
-import static it.infn.mw.iam.api.tokens.TokensControllerSupport.CONTENT_TYPE;
+import static it.infn.mw.iam.api.tokens.TokensControllerSupport.APPLICATION_JSON_CONTENT_TYPE;
 import static it.infn.mw.iam.api.tokens.TokensControllerSupport.TOKENS_MAX_PAGE_SIZE;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.google.common.collect.Lists;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import it.infn.mw.iam.IamLoginService;
-import it.infn.mw.iam.api.scim.converter.ScimResourceLocationProvider;
-import it.infn.mw.iam.api.tokens.model.AccessToken;
-import it.infn.mw.iam.api.tokens.model.TokensListResponse;
-import it.infn.mw.iam.persistence.model.IamAccount;
-import it.infn.mw.iam.persistence.repository.IamOAuthAccessTokenRepository;
-import it.infn.mw.iam.test.core.CoreControllerTestSupport;
-import it.infn.mw.iam.test.util.WithMockOAuthUser;
+import java.util.List;
 
 import org.junit.After;
 import org.junit.Before;
@@ -31,7 +20,20 @@ import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
-import java.util.List;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Lists;
+
+import it.infn.mw.iam.IamLoginService;
+import it.infn.mw.iam.api.scim.converter.ScimResourceLocationProvider;
+import it.infn.mw.iam.api.tokens.model.AccessToken;
+import it.infn.mw.iam.api.tokens.model.TokensListResponse;
+import it.infn.mw.iam.persistence.model.IamAccount;
+import it.infn.mw.iam.persistence.repository.IamOAuthAccessTokenRepository;
+import it.infn.mw.iam.test.core.CoreControllerTestSupport;
+import it.infn.mw.iam.test.util.DateEqualModulo1Second;
+import it.infn.mw.iam.test.util.WithMockOAuthUser;
+import it.infn.mw.iam.test.util.oauth.MockOAuth2Filter;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = {IamLoginService.class, CoreControllerTestSupport.class})
@@ -57,27 +59,32 @@ public class AccessTokenGetListTests extends TestTokensUtils {
   private static final String TESTUSER_USERNAME = "test_102";
   private static final String TESTUSER2_USERNAME = "test_103";
   private static final String PARTIAL_USERNAME = "test_10";
+  
+  @Autowired
+  private MockOAuth2Filter mockOAuth2Filter;
 
   @Before
   public void setup() {
     clearAllTokens();
+    mockOAuth2Filter.cleanupSecurityContext();
     initMvc();
   }
 
   @After
   public void teardown() {
     clearAllTokens();
+    mockOAuth2Filter.cleanupSecurityContext();
   }
 
   @Test
   public void getEmptyAccessTokenList() throws Exception {
 
     TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
+        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE))
+          .andExpect(status().isOk())
+          .andReturn()
+          .getResponse()
+          .getContentAsString(),
         new TypeReference<TokensListResponse<AccessToken>>() {});
 
     assertThat(atl.getTotalResults(), equalTo(0L));
@@ -90,15 +97,15 @@ public class AccessTokenGetListTests extends TestTokensUtils {
 
     buildAccessToken(loadTestClient(TEST_CLIENT_ID), TESTUSER_USERNAME, SCOPES);
 
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE).param("count", "0"))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE)
+        .param("count", "0"))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(2L));
+    assertThat(atl.getTotalResults(), equalTo(1L));
     assertThat(atl.getStartIndex(), equalTo(1L));
     assertThat(atl.getItemsPerPage(), equalTo(0L));
   }
@@ -111,27 +118,26 @@ public class AccessTokenGetListTests extends TestTokensUtils {
 
     OAuth2AccessTokenEntity at = buildAccessToken(client, TESTUSER_USERNAME, SCOPES);
 
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE).param("attributes",
-            "user,idToken"))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE)
+        .param("attributes", "user,idToken"))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(2L));
+    assertThat(atl.getTotalResults(), equalTo(1L));
     assertThat(atl.getStartIndex(), equalTo(1L));
-    assertThat(atl.getItemsPerPage(), equalTo(2L));
-    assertThat(atl.getResources().size(), equalTo(2));
+    assertThat(atl.getItemsPerPage(), equalTo(1L));
+    assertThat(atl.getResources().size(), equalTo(1));
 
     List<AccessToken> acl = atl.getResources();
-    AccessToken remoteAt = acl.stream().filter(a -> a.getIdToken() != null).findFirst().get();
+    AccessToken remoteAt = acl.get(0);
 
     assertThat(remoteAt.getId(), equalTo(at.getId()));
     assertThat(remoteAt.getClient(), equalTo(null));
     assertThat(remoteAt.getExpiration(), equalTo(null));
-    assertThat(remoteAt.getIdToken().getId(), equalTo(at.getIdToken().getId()));
+
     assertThat(remoteAt.getUser().getId(), equalTo(user.getUuid()));
     assertThat(remoteAt.getUser().getUserName(), equalTo(user.getUsername()));
     assertThat(remoteAt.getUser().getRef(),
@@ -151,30 +157,29 @@ public class AccessTokenGetListTests extends TestTokensUtils {
     accessTokens.add(target);
     accessTokens.add(buildAccessToken(client2, TESTUSER_USERNAME, SCOPES));
 
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE).param("clientId",
-            client1.getClientId()))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE)
+        .param("clientId", client1.getClientId()))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(2L));
+    assertThat(atl.getTotalResults(), equalTo(1L));
     assertThat(atl.getStartIndex(), equalTo(1L));
-    assertThat(atl.getItemsPerPage(), equalTo(2L));
-    assertThat(atl.getResources().size(), equalTo(2));
+    assertThat(atl.getItemsPerPage(), equalTo(1L));
+    assertThat(atl.getResources().size(), equalTo(1));
 
     List<AccessToken> acl = atl.getResources();
-    AccessToken remoteAt = acl.stream().filter(a -> a.getIdToken() != null).findFirst().get();
+    AccessToken remoteAt = acl.get(0);
 
     assertThat(remoteAt.getId(), equalTo(target.getId()));
     assertThat(remoteAt.getClient().getId(), equalTo(target.getClient().getId()));
     assertThat(remoteAt.getClient().getClientId(), equalTo(target.getClient().getClientId()));
     assertThat(remoteAt.getClient().getRef(), equalTo(target.getClient().getClientUri()));
 
-    assertThat(remoteAt.getExpiration(), equalTo(target.getExpiration()));
-    assertThat(remoteAt.getIdToken().getId(), equalTo(target.getIdToken().getId()));
+    assertThat(remoteAt.getExpiration(), new DateEqualModulo1Second(target.getExpiration()));
+
     assertThat(remoteAt.getUser().getId(), equalTo(user.getUuid()));
     assertThat(remoteAt.getUser().getUserName(), equalTo(user.getUsername()));
     assertThat(remoteAt.getUser().getRef(),
@@ -193,30 +198,29 @@ public class AccessTokenGetListTests extends TestTokensUtils {
     accessTokens.add(target);
     accessTokens.add(buildAccessToken(client, TESTUSER2_USERNAME, SCOPES));
 
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE).param("userId",
-            user1.getUsername()))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE)
+        .param("userId", user1.getUsername()))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(2L));
+    assertThat(atl.getTotalResults(), equalTo(1L));
     assertThat(atl.getStartIndex(), equalTo(1L));
-    assertThat(atl.getItemsPerPage(), equalTo(2L));
-    assertThat(atl.getResources().size(), equalTo(2));
+    assertThat(atl.getItemsPerPage(), equalTo(1L));
+    assertThat(atl.getResources().size(), equalTo(1));
 
     List<AccessToken> acl = atl.getResources();
-    AccessToken remoteAt = acl.stream().filter(a -> a.getIdToken() != null).findFirst().get();
+    AccessToken remoteAt = acl.get(0);
 
     assertThat(remoteAt.getId(), equalTo(target.getId()));
     assertThat(remoteAt.getClient().getId(), equalTo(target.getClient().getId()));
     assertThat(remoteAt.getClient().getClientId(), equalTo(target.getClient().getClientId()));
     assertThat(remoteAt.getClient().getRef(), equalTo(target.getClient().getClientUri()));
 
-    assertThat(remoteAt.getExpiration(), equalTo(target.getExpiration()));
-    assertThat(remoteAt.getIdToken().getId(), equalTo(target.getIdToken().getId()));
+    assertThat(remoteAt.getExpiration(), new DateEqualModulo1Second(target.getExpiration()));
+
     assertThat(remoteAt.getUser().getId(), equalTo(user1.getUuid()));
     assertThat(remoteAt.getUser().getUserName(), equalTo(user1.getUsername()));
     assertThat(remoteAt.getUser().getRef(),
@@ -238,31 +242,30 @@ public class AccessTokenGetListTests extends TestTokensUtils {
     accessTokens.add(buildAccessToken(client2, TESTUSER_USERNAME, SCOPES));
     accessTokens.add(buildAccessToken(client2, TESTUSER2_USERNAME, SCOPES));
 
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE)
-            .param("userId", user1.getUsername())
-            .param("clientId", client1.getClientId()))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE)
+        .param("userId", user1.getUsername())
+        .param("clientId", client1.getClientId()))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(2L));
+    assertThat(atl.getTotalResults(), equalTo(1L));
     assertThat(atl.getStartIndex(), equalTo(1L));
-    assertThat(atl.getItemsPerPage(), equalTo(2L));
-    assertThat(atl.getResources().size(), equalTo(2));
+    assertThat(atl.getItemsPerPage(), equalTo(1L));
+    assertThat(atl.getResources().size(), equalTo(1));
 
     List<AccessToken> acl = atl.getResources();
-    AccessToken remoteAt = acl.stream().filter(a -> a.getIdToken() != null).findFirst().get();
+    AccessToken remoteAt = acl.get(0);
 
     assertThat(remoteAt.getId(), equalTo(target.getId()));
     assertThat(remoteAt.getClient().getId(), equalTo(target.getClient().getId()));
     assertThat(remoteAt.getClient().getClientId(), equalTo(target.getClient().getClientId()));
     assertThat(remoteAt.getClient().getRef(), equalTo(target.getClient().getClientUri()));
 
-    assertThat(remoteAt.getExpiration(), equalTo(target.getExpiration()));
-    assertThat(remoteAt.getIdToken().getId(), equalTo(target.getIdToken().getId()));
+    assertThat(remoteAt.getExpiration(), new DateEqualModulo1Second(target.getExpiration()));
+    
     assertThat(remoteAt.getUser().getId(), equalTo(user1.getUuid()));
     assertThat(remoteAt.getUser().getUserName(), equalTo(user1.getUsername()));
     assertThat(remoteAt.getUser().getRef(),
@@ -277,14 +280,13 @@ public class AccessTokenGetListTests extends TestTokensUtils {
     buildAccessToken(client, TESTUSER_USERNAME, SCOPES);
     buildAccessToken(client, TESTUSER2_USERNAME, SCOPES);
 
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE).param("userId",
-            PARTIAL_USERNAME))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE)
+        .param("userId", PARTIAL_USERNAME))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
     assertThat(atl.getTotalResults(), equalTo(0L));
     assertThat(atl.getStartIndex(), equalTo(1L));
@@ -301,14 +303,14 @@ public class AccessTokenGetListTests extends TestTokensUtils {
 
     /* get first page */
     TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
+        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE))
+          .andExpect(status().isOk())
+          .andReturn()
+          .getResponse()
+          .getContentAsString(),
         new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(2L * TOKENS_MAX_PAGE_SIZE));
+    assertThat(atl.getTotalResults(), equalTo(Long.valueOf(TOKENS_MAX_PAGE_SIZE)));
     assertThat(atl.getStartIndex(), equalTo(1L));
     assertThat(atl.getItemsPerPage(), equalTo(Long.valueOf(TOKENS_MAX_PAGE_SIZE)));
     assertThat(atl.getResources().size(), equalTo(TOKENS_MAX_PAGE_SIZE));
@@ -322,19 +324,18 @@ public class AccessTokenGetListTests extends TestTokensUtils {
     }
 
     /* get second page */
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE).param("startIndex",
-            String.valueOf(TOKENS_MAX_PAGE_SIZE)))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE)
+        .param("startIndex", String.valueOf(TOKENS_MAX_PAGE_SIZE)))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(2L * TOKENS_MAX_PAGE_SIZE));
+    assertThat(atl.getTotalResults(), equalTo(Long.valueOf(TOKENS_MAX_PAGE_SIZE)));
     assertThat(atl.getStartIndex(), equalTo(Long.valueOf(TOKENS_MAX_PAGE_SIZE)));
-    assertThat(atl.getItemsPerPage(), equalTo(Long.valueOf(TOKENS_MAX_PAGE_SIZE)));
-    assertThat(atl.getResources().size(), equalTo(TOKENS_MAX_PAGE_SIZE));
+    assertThat(atl.getItemsPerPage(), equalTo(1L));
+    assertThat(atl.getResources().size(), equalTo(1));
   }
 
   @Test
@@ -344,23 +345,22 @@ public class AccessTokenGetListTests extends TestTokensUtils {
 
     buildAccessToken(client, TESTUSER_USERNAME, SCOPES);
 
-    assertThat(tokenRepository.count(), equalTo(2L));
+    assertThat(tokenRepository.count(), equalTo(1L));
 
-    TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE).param("userId",
-            "1%; DELETE FROM access_token; SELECT * FROM access_token WHERE userId LIKE %"))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
-        new TypeReference<TokensListResponse<AccessToken>>() {});
+    TokensListResponse<AccessToken> atl = mapper.readValue(mvc
+      .perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE).param(
+          "userId", "1%; DELETE FROM access_token; SELECT * FROM access_token WHERE userId LIKE %"))
+      .andExpect(status().isOk())
+      .andReturn()
+      .getResponse()
+      .getContentAsString(), new TypeReference<TokensListResponse<AccessToken>>() {});
 
     assertThat(atl.getTotalResults(), equalTo(0L));
     assertThat(atl.getStartIndex(), equalTo(1L));
     assertThat(atl.getItemsPerPage(), equalTo(0L));
     assertThat(atl.getResources().size(), equalTo(0));
 
-    assertThat(tokenRepository.count(), equalTo(2L));
+    assertThat(tokenRepository.count(), equalTo(1L));
   }
 
 
@@ -371,15 +371,15 @@ public class AccessTokenGetListTests extends TestTokensUtils {
     buildAccessToken(loadTestClient(TEST_CLIENT_ID), SCOPES);
 
     TokensListResponse<AccessToken> atl = mapper.readValue(
-        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(CONTENT_TYPE))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString(),
+        mvc.perform(get(ACCESS_TOKENS_BASE_PATH).contentType(APPLICATION_JSON_CONTENT_TYPE))
+          .andExpect(status().isOk())
+          .andReturn()
+          .getResponse()
+          .getContentAsString(),
         new TypeReference<TokensListResponse<AccessToken>>() {});
 
-    assertThat(atl.getTotalResults(), equalTo(3L));
+    assertThat(atl.getTotalResults(), equalTo(2L));
     assertThat(atl.getStartIndex(), equalTo(1L));
-    assertThat(atl.getItemsPerPage(), equalTo(3L));
+    assertThat(atl.getItemsPerPage(), equalTo(2L));
   }
 }
